updating export files

+++ q100/models/qscope_ABM.gaml
@@ -384,10 +384,12 @@ global {
 
 	init {
 
-    bool delete_csv_export_emissions <- delete_file("../data/outputs/output/emissions/");
-    bool delete_csv_export_energy_prices <- delete_file("../data/outputs/output/energy_prices/");
-    bool delete_csv_export_connections <- delete_file("../data/outputs/output/connections/");
-
+	if (timestamp = "") // only delete files in general output folder if using GUI
+	{
+    	bool delete_csv_export_emissions <- delete_file("../data/outputs/output/emissions/");
+    	bool delete_csv_export_energy_prices <- delete_file("../data/outputs/output/energy_prices/");
+    	bool delete_csv_export_connections <- delete_file("../data/outputs/output/connections/");
+	}
 		create technical_data_calculator number: 1;
 		create building from: shape_file_buildings with: [id::string(read("Kataster_C")), type::string(read(attributes_source)), units::int(read("Kataster_W")), street::string(read("Kataster_S")), mod_status::string(read("Kataster_8")), net_floor_area::int(read("Kataster_6")), spec_heat_consumption::float(read("Kataster13")), spec_power_consumption::float(read("Kataster15")), energy_source::string(read("Kataster_E"))] { // create agents according to shapefile metadata
 
@@ -1869,7 +1871,7 @@ experiment agent_decision_making type: gui{
 
 	reflex save_energy_costs {
 		if (current_date.month = 1) and (current_date.day = 1) {
-				string export_file <- (timestamp = "") ? "../data/outputs/output/energy_prices/energy_prices_export.csv" : "../data/outputs/output_" + timestamp + "/energy_prices/energy_prices_export.csv";
+				string export_file <- (timestamp = "") ? "../data/outputs/output/energy_prices/energy_prices_total.csv" : "../data/outputs/output_" + timestamp + "/energy_prices/energy_prices_total.csv";
 
 			save [cycle, current_date, power_price, oil_price, gas_price]
 			to: export_file type: csv rewrite: false  header: true;
